+++ ssr/src/page/confirm_booking/confirmation.rs
@@ -1,6 +1,6 @@
 use crate::{
     page::{
-        confirmation::payment_handler::PaymentHandler, BookRoomHandler, BookingHandler, Navbar,
+        confirm_booking::payment_handler::PaymentHandler, BookRoomHandler, BookingHandler, Navbar,
     },
     state::{search_state::SearchCtx, view_state::HotelInfoCtx},
 };
