+++ ssr/src/page/confirm_booking/mod.rs
